Add Networking
- Add Standalone Asio as Engine Dependency
- Remove Standalone freetype Engine Dependency
- Add numerous networking functions to project settings
- Initialize Audio and Scripting system in OnAttach() instead of Application constructor
- Initialize networking context for OnSceneStart/Stop() in editor and runtime
- Add .env file for projects to manage server settings
- Serialize/Deserialize .env file through project loading setup in AssetManager
- Refactor Application Run() to allow for constant delta time in physics simulation and include std::chrono library usage
- Add API for skipping and adding additional update cycles in main/update thread
- Add ProcessEventQueue() to Application and Networking contexts to handle async events
- Add AsyncBusyTimer in Engine Core to allow for expensive and precise timers that run on their own thread
- Added numerous event types to handle networking code and async timer code
- Added client/server networking code into Engine/Networking
- Add network component stubs for later development in Scene/Components.h
- Add functionality inside RuntimeUI to select widget, set widget text color, set widget background color, and make a widget selectable.
- Update Pong to include Online Gameplay mode
- Add Server App for hosting
- Add OnPhysicsCollisionEnd event/pipeline for scripting
- Update Logging to use trace as default debug logger
- Replace KG_CORE_*** with KG_*** when logging
- Add CRC32 hashing functionality in Core/Filesystem.h
- Update AppTick Generator API to dynamically add/remove generators when needed
- Add Recurring Timer to AsyncBusyTimer API

+++ Engine/dependencies/msdf_atlas_gen/msdfgen/freetype/src/psaux/psintrp.c
@@ -1002,7 +1002,7 @@
 
 
             if ( val )
-              subrNum = *val;
+              subrNum = (int)*val;
             else
               subrNum = -1;
           }
