+++ Engine/dependencies/msdf_atlas_gen/msdfgen/freetype/src/psaux/t1decode.c
@@ -1990,7 +1990,7 @@
 
 
               if ( val )
-                idx = *val;
+              idx = (int)*val;
               else
                 idx = -1;
             }
