+++ Engine/dependencies/msdf_atlas_gen/msdfgen/freetype/src/smooth/ftgrays.c
@@ -1968,7 +1968,7 @@ typedef ptrdiff_t  FT_PtrDist;
     for ( y = yMin; y < yMax; )
     {
       ras.min_ey = y;
-      y         += height;
+      y         += (int)height;
       ras.max_ey = FT_MIN( y, yMax );
 
       band    = bands;
